Adds loading in model and selecting cluster based on game concepts
This commit includes the jar for the pmml library, a KMeans model.pmml (example) and the code to load this model in our AI, get the cluster of a game based on this model, and pick an AI based on this cluster. The AI chosen right now is arbitrary, and with this model the second cluster is always picked. So its more the setup for the actual model.